+++ test/utils/project_utils_test.dart
@@ -112,14 +112,14 @@ description: test
     });
 
     test('primitiveValue should return primitive defaults', () {
-      expect(projectUtil.primitiveValue('int'), '1');
-      expect(projectUtil.primitiveValue('double'), '1.0');
-      expect(projectUtil.primitiveValue('bool'), 'true');
-      expect(projectUtil.primitiveValue('String'), "'test'");
+      expect(projectUtil.primitiveValueForAssert('int'), '1');
+      expect(projectUtil.primitiveValueForAssert('double'), '1.0');
+      expect(projectUtil.primitiveValueForAssert('bool'), 'true');
+      expect(projectUtil.primitiveValueForAssert('String'), "'test'");
     });
 
     test('primitiveValue should return null for unknown types', () {
-      expect(projectUtil.primitiveValue('UserService'), 'null');
+      expect(projectUtil.primitiveValueForAssert('UserService'), 'null');
     });
   });
 }